Separate function for toggling 'completed' attribute
