+++ src/index.js
@@ -10,6 +10,8 @@ import Interface from './modules/interface.js';
 
 import removeFromStorage from './modules/removeFromStorage.js';
 
+import toggleCompleted from './modules/toggleCompleted.js';
+
 document.addEventListener('DOMContentLoaded', Interface.displayTasks);
 
 document.querySelector('.form').addEventListener('submit', (e) => {
@@ -57,19 +59,7 @@ document.querySelector('#toDoItems').addEventListener('keypress', (e) => {
   }
 });
 
-document.querySelector('#toDoItems').addEventListener('change', (e) => {
-  let tasks = [];
-  if (localStorage.getItem('tasks') !== null) {
-    tasks = JSON.parse(localStorage.getItem('tasks'));
-  }
-  const { id } = e.target.parentElement.parentElement;
-  if (e.target.checked) {
-    tasks[id - 1].completed = true;
-  } else {
-    tasks[id - 1].completed = false;
-  }
-  localStorage.setItem('tasks', JSON.stringify(tasks));
-});
+document.querySelector('#toDoItems').addEventListener('change', toggleCompleted);
 
 document.querySelector('#clearAllBtn').addEventListener('click', () => {
   let tasks = [];
